+++ libvips/morphology/nearest.c
@@ -116,7 +116,7 @@ vips_fill_nearest_pixel(Circle *circle, int x, int y, int octant)
 	p = (float *) VIPS_IMAGE_ADDR(circle->nearest->distance, x, y);
 	dx = x - circle->seed->x;
 	dy = y - circle->seed->y;
-	radius = sqrt(dx * dx + dy * dy);
+	radius = sqrtf(dx * dx + dy * dy);
 
 	if (p[0] == 0 ||
 		p[0] > radius) {
